update integration to new db model type

+++ src/WindLordApi.Integrations/MetYr/MetYrDto.cs
@@ -7,26 +7,26 @@ namespace WindLordApi.Integrations.MetYr;
 public record MetYrDto
 {
     public required string Time { get; init; }
-    public required double AirPressureAtSeaLevel { get; init; }
-    public required double AirTemperature { get; init; }
-    public double? AirTemperaturePercentile10 { get; init; }
-    public double? AirTemperaturePercentile90 { get; init; }
+    public required decimal AirPressureAtSeaLevel { get; init; }
+    public required decimal AirTemperature { get; init; }
+    public decimal? AirTemperaturePercentile10 { get; init; }
+    public decimal? AirTemperaturePercentile90 { get; init; }
     public required double CloudAreaFraction { get; init; }
     public required double CloudAreaFractionHigh { get; init; }
     public required double CloudAreaFractionLow { get; init; }
     public required double CloudAreaFractionMedium { get; init; }
-    public required double DewPointTemperature { get; init; }
+    public required decimal DewPointTemperature { get; init; }
     public required double RelativeHumidity { get; init; }
     public required double WindFromDirection { get; init; }
-    public required double WindSpeed { get; init; }
-    public required double PrecipitationAmount { get; init; }
+    public required decimal WindSpeed { get; init; }
+    public required decimal PrecipitationAmount { get; init; }
     public double? PrecipitationAmountMax { get; init; }
     public double? PrecipitationAmountMin { get; init; }
-    public double? ProbabilityOfPrecipitation { get; init; }
+    public float? ProbabilityOfPrecipitation { get; init; }
     public required string SymbolCode { get; init; }
     public required double FogAreaFraction { get; init; }
     public required double UltravioletIndexClearSky { get; init; }
-    public double? WindSpeedOfGust { get; init; }
+    public decimal? WindSpeedOfGust { get; init; }
     public double? ProbabilityOfThunder { get; init; }
     public required string Next6HoursSymbolCode { get; init; }
 }
